:card_file_box: Corrige las migraciones del nivel 4

+++ database/migrations/2022_08_01_033726_create_personas_table.php
@@ -25,7 +25,7 @@ return new class extends Migration {
             $table->foreignId('distrito_id')->nullable()->constrained('distritos')
                 ->cascadeOnUpdate()
                 ->nullOnDelete();
-            $table->smallInteger('sexo_id');
+            $table->tinyInteger('sexo_id'); // -127 ~ 127
             $table->timestamps();
         });
     }
